--- app/__init__.py
@@ -1,9 +1,0 @@
-from flask import Flask
-
-
-# Initializing application
-app = Flask(__name__)
-
-
-
-from app import models
