--- app/templates/about.html
@@ -1,13 +1,0 @@
-{% extends 'base.html'%}
-
-{% block navbar %}
-    {% include 'base.html' %}
-{% endblock %}
-
-
-{% block content %}
-this is the about section
-      
-      
-
-{% endblock %}
