--- app/templates/index.html
@@ -1,11 +1,0 @@
-{% extends 'base.html'%}
-
-{% block navbar %}
-    {% include 'base.html' %}
-{% endblock %}
-
-{% block content %}
-     homepage 
-      
-
-{% endblock %}
