--- app/templates/men.html
@@ -1,23 +1,0 @@
-{% extends 'base.html'%}
-
-{% block navbar %}
-    {% include 'base.html' %}
-{% endblock %}
-
-{% block content %}
-<div class="container">
-      
-<div class="card" style="width: 18rem;">
-    <img class="card-img-top" src="app/static/images/featured1.png" alt="">
-    <div class="card-body">
-      <h5 class="card-title">Card title</h5>
-      
-      <a href="#" class="btn btn-primary">Add to cart</a>
-    </div>
-  </div>
-
-</div>
-  
-
-{% endblock %}
-
