+++ static/CSS/style.css
@@ -1,0 +1,7 @@
+body , html {
+background-color: antiquewhite;
+}
+
+
+
+
