+++ templates/about.html
@@ -1,0 +1,28 @@
+{% extends 'base.html'%}
+
+{% block navbar %}
+    {% include 'base.html' %}
+{% endblock %}
+
+{% block content %}
+     about
+      
+
+{% endblock %}
+
+<br><br>
+    <!-- <div class="container">
+        <div class="row">
+          <div class="col">
+            <img src="static/images/about us.jpeg"  height="400px" width="500px"  class="img-thumbnail">
+          </div>
+          <div class="col">
+            <h5>Pizza Delicious was founder in 1945. Our Motto is <i>family, friends, and pizza</i>.</h5>
+            <h5>We use our art to bring friends and families together with our yummy pizza.</h5>
+          </div>
+        </div>
+      </div> -->
+
+ 
+
+
