+++ templates/base.html
@@ -8,7 +8,8 @@
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <!-- css link -->
-    <link rel="stylesheet" href="app/static/CSS/style.css">
+    <link rel="stylesheet" href="static/CSS/style.css">
+    <link  rel="stylesheet" href="{{ url_for('static', filename='CSS/style.css') }}" />
     <title>ShoeHub</title>
   
   </head>
@@ -16,7 +17,7 @@
       
 
 {% block navbar %}
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
+<nav class="navbar navbar-expand-lg navbar-light bg-primary">
     <div class="container-fluid">
       <a class="navbar-brand" href="#">ShoeHub</a>
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
@@ -40,8 +41,6 @@
           <li class="nav-item">
             <a class="nav-link active" aria-current="page" href="/contact">Contact Us</a>
           </li>
-
-          
         </ul>
         
       </div>
