+++ templates/contact.html
@@ -1,0 +1,38 @@
+{% extends 'base.html'%}
+
+{% block navbar %}
+    {% include 'base.html' %}
+{% endblock %}
+
+{% block content %}
+  contact
+      
+
+{% endblock %}
+        
+        <!-- <div class="container">
+            <form action="" method="POST">
+                <div class="form-group">
+                  <label for="exampleInputEmail1">First Name...</label>
+                  <input type="text" class="form-control" placeholder="First Name" name ="first_name"  >
+                 
+                </div>
+                <div class="form-group">
+                    <label for="exampleInputEmail1">Last Name...</label>
+                    <input type="text" class="form-control"  placeholder="Last Name" name ="last_name" >
+                   
+                  </div>
+                  <div class="form-group">
+                    <label for="exampleInputEmail1">Email address...</label>
+                    <input type="email" class="form-control"  placeholder="Email Adress" name ="email" >
+                   
+                  </div>
+                <br><br>
+                <button type="submit" class="btn btn-primary">Submit</button>
+              </form>
+            </div> -->
+  
+    
+
+    
+
