+++ templates/index.html
@@ -1,0 +1,34 @@
+{% extends 'base.html'%}
+
+{% block navbar %}
+    {% include 'base.html' %}
+{% endblock %}
+
+{% block content %}
+<br><br>
+    <div class="container" >
+    <br><br><br>
+        <div class="row">
+          <div class="col-md">
+            <h3>Welcome to ShoeHub where you get quality shoe at at affordable price. </h3>
+            <br><br>
+            <button type="button" class="btn btn-primary">Login</button>
+            <button type="button" class="btn btn-primary">SignUp</button>
+          </div>
+          <div class="col-md">
+            <img src="static/images/boy1.jpeg" alt="" class="rounded-circle" height="500px">
+          </div>
+          
+        </div>
+        
+
+
+
+        
+    
+
+
+    </div>
+      
+
+{% endblock %}
