+++ templates/men.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
+  men
       
-   categories   
 
 {% endblock %}
