+++ templates/women.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
+     women
       
- category   
 
 {% endblock %}
